Refactor: Remove OAuth2Config, builder, and related setter method on EsiClient. Remove config directory under oauth2 & move client.rs under oauth2/cient.rs
