--- src/oauth2/config/builder.rs
@@ -1,508 +1,0 @@
-//! # EVE Online OAuth2 Config Builder
-//!
-//! Provides methods to override the default OAuth2 settings for the eve_esi crate. This allows the
-//! modification of OAuth2 endpoint URLs or the logic of how JWT (JSON web token) key caching and refreshing is
-//! handled which are used to validate tokens for user authentication.
-//!
-//! - See [module-level] documentation for a higher level overview and usage example
-//!
-//! ## Features
-//! - Override default authentication, JWT token, JWT token key endpoint URLs
-//! - Adjust expiration times for JWT key cache
-//! - Adjust the timeout for waiting for JWT key refreshes
-//! - Adjust wait time/backoff period beteween refresh and how many retries should be made to fetch JWT keys
-//! - Enable/disable the proactive background JWT key refresh
-//!
-//! ## Builder Methods
-//!
-//! | Method          | Purpose                                    |
-//! | --------------- | ------------------------------------------ |
-//! | `new`           | Create a new OAuth2Config builder          |
-//! | `build`         | Build the OAuth2 config                    |
-//! | `auth_url`      | URL for EVE OAuth2 authentication          |
-//! | `token_url`     | URL for EVE OAuth2 token                   |
-//! | `jwk_url`       | URL for EVE OAuth2 token keys              |
-//! | `jwk_cache_ttl` | The time that JWT keys are cached for      |
-//! | `jwk_refresh_max_retries` | Amount of retries when a key fetch fails |
-//! | `jwk_refresh_backoff`     | How long to wait between retries |
-//! | `jwk_refresh_timeout`     | How long to wait for another thread to refresh |
-//! | `jwk_refresh_cooldown`    | Cooldown between sets of JWT key refresh attempts |
-//! | `jwk_background_refresh_enabled` | Enable/disable background refresh |
-//! | `jwk_background_refresh_threshold_percent` | Percentage at which cache is refreshed proactively |
-
-use oauth2::{AuthUrl, TokenUrl};
-
-use crate::{
-    constant::{
-        DEFAULT_AUTH_URL, DEFAULT_JWK_BACKGROUND_REFRESH_THRESHOLD_PERCENT, DEFAULT_JWK_CACHE_TTL,
-        DEFAULT_JWK_REFRESH_BACKOFF, DEFAULT_JWK_REFRESH_COOLDOWN, DEFAULT_JWK_REFRESH_MAX_RETRIES,
-        DEFAULT_JWK_REFRESH_TIMEOUT, DEFAULT_JWK_URL, DEFAULT_TOKEN_URL,
-    },
-    error::EsiError,
-    oauth2::error::OAuthConfigError,
-};
-
-use super::config::OAuth2Config;
-
-/// Builder struct for configuring & constructing an [`OAuth2Config`]
-///
-/// For a full overview, features, and usage examples, see the [module-level documentation](self).
-pub struct OAuth2ConfigBuilder {
-    // EVE OAuth2 API URL overrides
-    /// Authentication URL endpoint for the EVE Online OAuth2 login flow
-    pub(crate) auth_url: String,
-    /// Token URL endpoint used to retrieve tokens to authenticate users
-    pub(crate) token_url: String,
-    /// JSON web token key URL that provides keys used to validate tokens
-    pub(crate) jwk_url: String,
-
-    // JWT key cache settings
-    /// JWT key cache lifetime before expiration in seconds (default 1 hour)
-    pub(crate) jwk_cache_ttl: u64,
-    /// Maximum number of retries for JWT key refresh when cache is empty or expired (default 2 retries)
-    pub(crate) jwk_refresh_max_retries: u64,
-    /// Backoff period in seconds after a JWT key refresh failure when cache is empty or expired (default 100 milliseconds)
-    pub(crate) jwk_refresh_backoff: u64,
-    /// Timeout in seconds when waiting for another thread to refresh JWT key (default 5 seconds)
-    pub(crate) jwk_refresh_timeout: u64,
-    /// Cooldown period in seconds after a failed set of JWT key refresh attempts (default 60 seconds)
-    pub(crate) jwk_refresh_cooldown: u64,
-
-    // JWT key cache background refresh settings
-    /// Determines whether or not a background task is spawned to refresh JWT keys nearing expiration proactively
-    pub(crate) jwk_background_refresh_enabled: bool,
-    /// Percentage of jwk_cache_ttl for when the background JWT key refresh is triggered (default 80%)
-    pub(crate) jwk_background_refresh_threshold_percent: u64,
-}
-
-impl OAuth2ConfigBuilder {
-    /// Creates a new [`OAuth2ConfigBuilder`] instance used to build an [`OAuth2Config`]
-    ///
-    /// For a full overview, features, and usage example, see the [module-level documentation](self).
-    ///
-    /// # Returns
-    /// - [`OAuth2ConfigBuilder`]: Instance with the default settings that can be overridden with setter methods.
-    pub fn new() -> Self {
-        Self {
-            // EVE OAuth2 API URL overrides
-            auth_url: DEFAULT_AUTH_URL.to_string(),
-            token_url: DEFAULT_TOKEN_URL.to_string(),
-            jwk_url: DEFAULT_JWK_URL.to_string(),
-
-            // JWT key cache settings
-            jwk_cache_ttl: DEFAULT_JWK_CACHE_TTL,
-            jwk_refresh_max_retries: DEFAULT_JWK_REFRESH_MAX_RETRIES,
-            jwk_refresh_backoff: DEFAULT_JWK_REFRESH_BACKOFF,
-            jwk_refresh_timeout: DEFAULT_JWK_REFRESH_TIMEOUT,
-            jwk_refresh_cooldown: DEFAULT_JWK_REFRESH_COOLDOWN,
-
-            // JWT key cache background refresh settings
-            jwk_background_refresh_enabled: true,
-            jwk_background_refresh_threshold_percent:
-                DEFAULT_JWK_BACKGROUND_REFRESH_THRESHOLD_PERCENT,
-        }
-    }
-
-    /// Builds an [`OAuth2Config`] instance
-    ///
-    /// Converts an [`OAuth2ConfigBuilder`] into an [`OAuth2`] config
-    /// with the configured values that were set with the builder methods.
-    ///
-    /// For a full overview, features, and usage example, see the [module-level documentation](self).
-    ///
-    /// # Returns
-    /// - [`OAuth2Config`]: instance with the settings configured on the builder
-    ///
-    /// # Errors
-    /// - [`EsiError`]: If jwk background refresh threshold percent, auth_url, or token_url is
-    ///   configured incorrectly.
-    pub fn build(self) -> Result<OAuth2Config, EsiError> {
-        // Ensure background refresh percentage is set properly
-        if !(self.jwk_background_refresh_threshold_percent > 0) {
-            return Err(EsiError::OAuthConfigError(
-                OAuthConfigError::InvalidBackgroundRefreshThreshold,
-            ));
-        }
-        if !(self.jwk_background_refresh_threshold_percent < 100) {
-            return Err(EsiError::OAuthConfigError(
-                OAuthConfigError::InvalidBackgroundRefreshThreshold,
-            ));
-        }
-
-        // OAuth2 Client URL overrides
-        let auth_url = match AuthUrl::new(self.auth_url) {
-            Ok(url) => url,
-            Err(_) => return Err(EsiError::OAuthConfigError(OAuthConfigError::InvalidAuthUrl)),
-        };
-        let token_url = match TokenUrl::new(self.token_url) {
-            Ok(url) => url,
-            Err(_) => {
-                return Err(EsiError::OAuthConfigError(
-                    OAuthConfigError::InvalidTokenUrl,
-                ))
-            }
-        };
-
-        // Return built config
-        Ok(OAuth2Config {
-            // EVE OAuth2 API URL overrides
-            auth_url: auth_url,
-            token_url: token_url,
-            jwk_url: self.jwk_url,
-
-            // JWT key cache settings
-            jwk_cache_ttl: self.jwk_cache_ttl,
-            jwk_refresh_max_retries: self.jwk_refresh_max_retries,
-            jwk_refresh_backoff: self.jwk_refresh_backoff,
-            jwk_refresh_timeout: self.jwk_refresh_timeout,
-            jwk_refresh_cooldown: self.jwk_refresh_cooldown,
-
-            // JWT key cache background refresh settings
-            jwk_background_refresh_enabled: self.jwk_background_refresh_enabled,
-            jwk_background_refresh_threshold_percent: self.jwk_background_refresh_threshold_percent,
-        })
-    }
-
-    /// Sets the EVE Online oauth2 authorize URL to a custom URL.
-    ///
-    /// This method configures the authorize URL for EVE Online oauth2.
-    /// This is generally used for tests using a mock server with crates such as
-    /// [mockito](https://crates.io/crates/mockito) to avoid actual ESI API calls.
-    ///
-    /// # Arguments
-    /// - `auth_url` (&[`str`]): The EVE Online oauth2 authorize URL.
-    ///
-    /// # Returns
-    /// - [`OAuth2Config`]: Instance with updated EVE Online oauth2 authorize URL configuration.
-    pub fn auth_url(mut self, auth_url: &str) -> Self {
-        self.auth_url = auth_url.to_string();
-        self
-    }
-
-    /// Sets the EVE Online oauth2 token URL to a custom URL.
-    ///
-    /// This method configures the token URL for EVE Online oauth2 to a custom URL.
-    /// This is generally used for tests using a mock server with crates such as
-    /// [mockito](https://crates.io/crates/mockito) to avoid actual ESI API calls.
-    ///
-    /// # Arguments
-    /// - `token_url` (&[`str`]): The EVE Online oauth2 token URL.
-    ///
-    /// # Returns
-    /// - [`OAuth2Config`]: Instance with updated EVE Online oauth2 token URL configuration.
-    pub fn token_url(mut self, token_url: &str) -> Self {
-        self.token_url = token_url.to_string();
-        self
-    }
-
-    /// Sets the EVE Online JWK URI to a custom URL.
-    ///
-    /// This method configures the JWK URI for EVE Online OAuth2 to a custom URL.
-    /// This is generally used for tests using a mock server with crates such as
-    /// [mockito](https://crates.io/crates/mockito) to avoid actual ESI API calls.
-    ///
-    /// # Arguments
-    /// - `jwk_url` (&[`str`]): The EVE Online JWK URL.
-    ///
-    /// # Returns
-    /// - [`OAuth2Config`]: Instance with updated EVE Online JWK URI configuration.
-    pub fn jwk_url(mut self, jwk_url: &str) -> Self {
-        self.jwk_url = jwk_url.to_string();
-        self
-    }
-
-    /// Modifies the default lifetime of the JWT keys stored in cache
-    ///
-    /// By default, JWT keys are stored in cache for 3600 seconds (1 hour)
-    /// before they are considered expired and need to be refreshed.
-    ///
-    /// Additionally, JWT keys are proactively refreshed by a background
-    /// task at 80% expiration. You may wish to modify the
-    /// [`Self::jwk_background_refresh_threshold_percent`] or disable the
-    /// background refresh altogether with [`Self::jwk_background_refresh_enabled`].
-    ///
-    /// # Arguments
-    /// - `seconds` ([`u64`]): The lifetime in seconds of the JWT keys stored in the cache.
-    ///
-    /// # Returns
-    /// - [`OAuth2Config`]: Instance with the updated JWT key cache TTL
-    pub fn jwk_cache_ttl(mut self, seconds: u64) -> Self {
-        self.jwk_cache_ttl = seconds;
-        self
-    }
-
-    /// Modifies the max amount of refresh attempts when fetching JWT keys
-    ///
-    /// This determines how many attempts are made to refresh JWT keys when
-    /// the cache is empty or fully expired and it is imperative to refresh the
-    /// cache in order to validate tokens.
-    ///
-    /// Between each fetch attempt there is an exponential backoff of 100ms by default
-    /// which can be modified with the [`Self::jwk_refresh_backoff`] method.
-    ///
-    /// This does not affect the background JWT key refresh which is instead modified
-    /// with the [`Self::jwk_background_refresh_cooldown`] method.
-    ///
-    /// # Arguments
-    /// - `retry_attempts` ([`u64`]): The amount of retry attempts if a JWT key fetch fails.
-    ///
-    /// # Returns
-    /// - [`OAuth2Config`]: Instance with the updated JWK refresh max retries
-    pub fn jwk_refresh_max_retries(mut self, retry_attempts: u64) -> Self {
-        self.jwk_refresh_max_retries = retry_attempts;
-        self
-    }
-
-    /// Modifies the exponential backoff duration in milliseconds between JWT key fetch retry attempts
-    ///
-    /// The default behavior is a 100ms exponential backoff between each retry attempt
-    /// to refresh JWT keys when the cache is either empty or expired.
-    ///
-    /// For example: 100ms, 200ms, 400ms, etc
-    ///
-    /// The amount of retry attempts can be modified with the [`Self::jwk_refresh_max_retries`]
-    /// method.
-    ///
-    /// This does not affect the background JWT key refresh which is instead modified
-    /// with the [`Self::jwk_background_refresh_cooldown`] method.
-    ///
-    /// # Arguments
-    /// - `backoff_milliseconds` ([`u64`]): The exponential backoff duration in milliseconds between each attempt.
-    ///
-    /// # Returns
-    /// - [`OAuth2Config`]: Instance with the updated exponential backoff
-    pub fn jwk_refresh_backoff(mut self, backoff_milliseconds: u64) -> Self {
-        self.jwk_refresh_backoff = backoff_milliseconds;
-        self
-    }
-
-    /// Modifies the timeout waiting for another thread to perform a JWT key cache refresh
-    ///
-    /// This library uses a refresh lock shared between threads to indicate if a JWT key
-    /// cache refresh is already in progress on another thread. If the JWT key cache is
-    /// currently empty or expired and this refresh lock is in place, the current thread
-    /// will wait for a default of 5 seconds before timing out if the refresh takes too long.
-    ///
-    /// # Arguments
-    /// - `timeout_seconds` ([`u64`]): The timeout in seconds to wait for another thread to complete a
-    ///   JWT key refresh.
-    ///
-    /// # Returns
-    /// - [`OAuth2Config`]: Instance with the modified timeout setting.
-    pub fn jwk_refresh_timeout(mut self, timeout_seconds: u64) -> Self {
-        self.jwk_refresh_timeout = timeout_seconds;
-        self
-    }
-
-    /// Modifies the cooldown between sets of JWT key cache refresh attempts in the event of failure
-    ///
-    /// By default, when a set of JWT key cache refresh attempts fail there will be a cooldown of 60 seconds
-    /// between the next set of attempts to refresh JWT keys before expiration.
-    ///
-    /// # Arguments
-    /// - `cooldown_seconds` ([`u64`]): Cooldown in seconds between background JWT key cache refresh attempts.
-    ///
-    /// # Returns
-    /// - [`OAuth2Config`]: Instance with the modified background refresh cooldown.
-    pub fn jwk_refresh_cooldown(mut self, cooldown_seconds: u64) -> Self {
-        self.jwk_refresh_cooldown = cooldown_seconds;
-        self
-    }
-
-    /// Modifies whether or not the proactive background refresh when JWT keys are almost expired is enabled
-    ///
-    /// By default, when the JWT key cache is nearing expiration at around 80%, a background refresh task
-    /// will be spawned to proactively refresh the keys. This behavior is thread safe and a more detailed
-    /// description of how it works can be found at [`crate::oauth2::OAuth2Api::get_jwt_keys`].
-    ///
-    /// This functionality has been built with high volume applications in mind and will work for the
-    /// vast majority of production use cases. In the instance where you do want to have more control
-    /// over proactive JWT key refreshes consider disabling this and using a cron task to perform
-    /// a refresh instead with the [`crate::oauth2::OAuth2Api::fetch_and_update_cache`] method which will
-    /// update the cache regardless of expiration status.
-    ///
-    /// You can modify the % at which the proactive background refresh is triggered with the
-    /// [`Self::jwk_background_refresh_threshold_percent`] method.
-    ///
-    /// # Arguments
-    /// - `background_refresh_enabled` ([`bool`]): A bool indicating whether or not the background refresh
-    ///   is enabled.
-    ///
-    /// # Returns
-    /// - [`OAuth2Config`]: Instance with the modified status of whether or not the background refresh is enabled.
-    pub fn jwk_background_refresh_enabled(mut self, background_refresh_enabled: bool) -> Self {
-        self.jwk_background_refresh_enabled = background_refresh_enabled;
-        self
-    }
-
-    /// The % of JWT key cache lifetime for when the proactive background JWT key cache refresh is triggered
-    ///
-    /// By default, when the JWT key cache reaches 80% of the default 3600 second cache lifetime, a proactive
-    /// JWT key cache refresh will be triggered next time the [`crate::oauth2::OAuth2Api::get_jwt_keys`]
-    /// method is called.
-    ///
-    /// Should the attempt fail, there will be a 60 second cooldown between attempts which can be modified with
-    /// the [`Self::jwk_background_refresh_cooldown`] method.
-    ///
-    /// # Arguments
-    /// - `threshold_percent` ([`u64`]): A number representing the percentage of when the refresh should be triggered.
-    ///
-    /// # Returns
-    /// - [`OAuth2Config`]: Instance with the modified proactive background refresh threshold percentage.
-    pub fn jwk_background_refresh_threshold_percent(mut self, threshold_percentage: u64) -> Self {
-        self.jwk_background_refresh_threshold_percent = threshold_percentage;
-        self
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    /// Ensures that all setter methods for [`OAuth2ConfigBuilder`] work as expected
-    ///
-    /// Test Setup
-    /// - Create a new instance of [`OAuth2ConfigBuilder`] and modify each field
-    /// - Build the [`OAuth2ConfigBuilder`] returning an [`OAuth2Config`]
-    ///
-    /// Assertions
-    /// - Assert all URLs were set as expected
-    /// - Assert all JWT key cache settings were set as expected
-    /// - Assert all JWT key cache background refresh settings were set as expected
-    #[test]
-    fn test_config_setter_methods() {
-        let config = OAuth2ConfigBuilder::new()
-            // URL settings
-            .auth_url("https://example.com")
-            .token_url("https://example.com")
-            .jwk_url("https://example.com")
-            // JWT key cache settings
-            .jwk_cache_ttl(0)
-            .jwk_refresh_max_retries(0)
-            .jwk_refresh_backoff(0)
-            .jwk_refresh_timeout(0)
-            .jwk_refresh_cooldown(0)
-            // JWT key cache background refresh settings
-            .jwk_background_refresh_enabled(false)
-            .jwk_background_refresh_threshold_percent(1)
-            .build()
-            .expect("Failed to build OAuth2 config");
-
-        // Assert URLs were set
-        let auth_url = AuthUrl::new("https://example.com".to_string()).unwrap();
-        let token_url = TokenUrl::new("https://example.com".to_string()).unwrap();
-
-        assert_eq!(config.auth_url, auth_url);
-        assert_eq!(config.token_url, token_url);
-        assert_eq!(config.jwk_url, "https://example.com");
-
-        // Assert JWT key cache settings were set
-        assert_eq!(config.jwk_cache_ttl, 0);
-        assert_eq!(config.jwk_refresh_max_retries, 0);
-        assert_eq!(config.jwk_refresh_backoff, 0);
-        assert_eq!(config.jwk_refresh_timeout, 0);
-        assert_eq!(config.jwk_refresh_cooldown, 0);
-
-        // Assert JWT key cache background refresh settings were set
-        assert_eq!(config.jwk_background_refresh_enabled, false);
-        assert_eq!(config.jwk_background_refresh_threshold_percent, 1);
-    }
-
-    /// Expect an error setting the jwk background refresh threshold percent to 0
-    ///
-    /// # Test Setup
-    /// - Attempt to build an OAuth2 config with the jwk_background_refresh_percent to 0
-    ///
-    /// # Assertions
-    /// - Assert result is an error
-    /// - Assert error is of type OAuthConfigError::InvalidBackgroundRefreshThreshold
-    #[test]
-    fn test_invalid_background_refresh_threshold_0() {
-        // Create an OAuth2 config with invalid auth_url
-        let result = OAuth2Config::builder()
-            .jwk_background_refresh_threshold_percent(0)
-            .build();
-
-        // Assert result is error
-        assert!(result.is_err());
-
-        // Assert error is of type OAuthConfigError::InvalidBackgroundRefreshThreshold
-        assert!(matches!(
-            result,
-            Err(EsiError::OAuthConfigError(
-                OAuthConfigError::InvalidBackgroundRefreshThreshold
-            ))
-        ))
-    }
-
-    /// Expect an error setting the jwk background refresh threshold percent to 100
-    ///
-    /// # Test Setup
-    /// - Attempt to build an OAuth2 config with the jwk_background_refresh_percent to 100
-    ///
-    /// # Assertions
-    /// - Assert result is an error
-    /// - Assert error is of type OAuthConfigError::InvalidBackgroundRefreshThreshold
-    #[test]
-    fn test_invalid_background_refresh_threshold_100() {
-        // Create an OAuth2 config with invalid auth_url
-        let result = OAuth2Config::builder()
-            .jwk_background_refresh_threshold_percent(100)
-            .build();
-
-        // Assert result is error
-        assert!(result.is_err());
-
-        // Assert error is of type OAuthConfigError::InvalidBackgroundRefreshThreshold
-        assert!(matches!(
-            result,
-            Err(EsiError::OAuthConfigError(
-                OAuthConfigError::InvalidBackgroundRefreshThreshold
-            ))
-        ))
-    }
-
-    /// Tests the attempting initialize an EsiClient for oauth2 with an invalid auth_url
-    ///
-    /// # Test Setup
-    /// - Attempt to build an OAuth2 config with the auth_url set to an invalid URL.
-    ///
-    /// # Assertions
-    /// - Verifies that the error response is EsiError::OAuthError(OAuthError::InvalidAuthUrl)
-    #[test]
-    fn test_invalid_auth_url() {
-        // Create an OAuth2 config with invalid auth_url
-        let result = OAuth2Config::builder().auth_url("invalid_url").build();
-
-        // Assert result is an Error
-        assert!(result.is_err());
-
-        match result {
-            // Assert error is of the OAuthConfigError:InvalidAuthUrl variant
-            Err(EsiError::OAuthConfigError(OAuthConfigError::InvalidAuthUrl)) => {}
-            _ => panic!("Expected InvalidAuthUrl error"),
-        }
-    }
-
-    /// Tests the attempting initialize an EsiClient for oauth2 with an invalid token_url
-    ///
-    /// # Test Setup
-    /// - Attempt to build an OAuth2 config with the token_url set to an invalid URL.
-    ///
-    /// # Assertions
-    /// - Verifies that the error response is EsiError::OAuthError(OAuthError::InvalidTokenUrl)
-    #[test]
-    fn test_invalid_token_url() {
-        // Create an OAuth2 config with invalid token_url
-        let result = OAuth2Config::builder().token_url("invalid_url").build();
-
-        // Assert result is an Error
-        assert!(result.is_err());
-
-        match result {
-            // Assert error is of the OAuthConfigError:InvalidTokenUrl variant
-            Err(EsiError::OAuthConfigError(OAuthConfigError::InvalidTokenUrl)) => {}
-            _ => panic!("Expected InvalidTokenUrl error"),
-        }
-    }
-}
